@@ -1,6 +1,6 @@
 {
   "name": "simsage",
-  "version": "5.0.16",
+  "version": "5.0.18",
   "description": "Simsage: breathe life into your data",
   "repository": "git@gitlab.com:peter3125/simsage.git",
   "homepage": "/",
@@ -19,56 +19,57 @@
   "devDependencies": {
     "babel-cli": "^6.26.0",
     "babel-core": "^6.26.3",
-    "babel-loader": "^7.1.2",
+    "babel-loader": "^7.1.5",
     "babel-polyfill": "^6.26.0",
     "babel-preset-env": "^1.7.0",
-    "babel-preset-es2015": "^6.24.1",
     "babel-preset-react": "^6.24.1",
     "babel-preset-stage-0": "^6.24.1",
-    "compression-webpack-plugin": "^1.1.12",
-    "copy-webpack-plugin": "^4.5.4",
+    "clean-webpack-plugin": "^3.0.0",
+    "compression-webpack-plugin": "^3.0.0",
+    "copy-webpack-plugin": "^5.0.4",
     "cross-env": "^5.2.0",
-    "css-loader": "^0.28.11",
-    "d3": "^5.7.0",
+    "css-loader": "^3.2.0",
+    "d3": "^5.11.0",
     "extract-text-webpack-plugin": "^4.0.0-beta.0",
     "html-loader": "^0.5.5",
     "html-webpack-plugin": "^3.2.0",
-    "path": "0.12.7",
-    "style-loader": "^0.23.1",
-    "uglifyjs-webpack-plugin": "^2.0.1",
-    "webpack": "^4.30.0",
-    "webpack-cli": "^3.3.2",
-    "webpack-dev-server": "^3.1.14"
+    "path": "^0.12.7",
+    "prettier": "^1.18.2",
+    "style-loader": "^1.0.0",
+    "uglifyjs-webpack-plugin": "^2.2.0",
+    "webpack": "^4.39.2",
+    "webpack-cli": "^3.3.7",
+    "webpack-dev-server": "^3.8.0"
   },
   "dependencies": {
-    "@babel/polyfill": "^7.0.0",
-    "@date-io/date-fns": "1.1.0",
-    "@material-ui/core": "^3.9.2",
-    "@material-ui/icons": "^3.0.1",
-    "axios": "^0.18.0",
-    "chart.js": "^2.7.3",
-    "cleave.js": "^1.4.7",
-    "date-fns": "^2.0.0-alpha.27",
-    "downshift": "^3.2.7",
-    "luxon": "^1.4.4",
-    "material-ui-pickers": "^2.0.1",
-    "moment": "^2.22.2",
-    "npm": "^6.10.2",
-    "query-string": "^5.0.0",
-    "react": "^16.8.6",
-    "react-app-polyfill": "^0.2.1",
-    "react-bootstrap": "^1.0.0-beta.8",
+    "@date-io/core": "^1.3.9",
+    "@date-io/date-fns": "^1.3.9",
+    "@material-ui/core": "^4.3.3",
+    "@material-ui/icons": "^4.2.1",
+    "axios": "^0.19.0",
+    "chart.js": "^2.8.0",
+    "cleave.js": "^1.5.3",
+    "core-js": "^3.2.1",
+    "downshift": "^3.2.12",
+    "luxon": "^1.17.2",
+    "moment": "^2.24.0",
+    "npm": "^6.11.2",
+    "query-string": "^6.8.2",
+    "react": "^16.9.0",
+    "react-app-polyfill": "^1.0.2",
     "react-chartjs-2": "^2.7.6",
-    "react-dom": "^16.8.6",
-    "react-global-configuration": "^1.3.0",
-    "react-redux": "^7.0.3",
-    "react-router": "^5.0.0",
-    "react-router-dom": "^5.0.0",
+    "react-datepicker": "^2.9.6",
+    "react-dom": "^16.9.0",
+    "react-redux": "^7.1.0",
+    "react-router": "^5.0.1",
+    "react-router-dom": "^5.0.1",
     "react-router-redux": "^4.0.8",
-    "redux": "^4.0.1",
+    "redux": "^4.0.4",
+    "redux-logger": "^3.0.6",
     "redux-persist": "^5.10.0",
     "redux-thunk": "^2.3.0",
-    "typeface-roboto": "0.0.54",
-    "url-loader": "^1.1.2"
+    "style-loader": "^1.0.0",
+    "typeface-roboto": "^0.0.75",
+    "url-loader": "^2.1.0"
   }
 }
