@@ -64,7 +64,7 @@ export class ProgramConvert extends React.Component {
         };
     }
 
-    componentWillReceiveProps(nextProps) {
+    UNSAFE_componentWillReceiveProps(nextProps) {
         if (nextProps) {
             this.setState({filter: nextProps.filter, kbId: nextProps.kbId});
         }
