@@ -41,7 +41,7 @@ export class SemanticEdit extends Component {
             this.state.onSave(null);
         }
     }
-    componentWillReceiveProps(props) {
+    UNSAFE_componentWillReceiveProps(props) {
         this.setState({
             open: props.open,
             semantic: props.semantic,
@@ -67,7 +67,7 @@ export class SemanticEdit extends Component {
 
                 <DialogTitle id="alert-dialog-title">Edit Semantic</DialogTitle>
                 <DialogContent>
-                    <Grid container spacing={8}>
+                    <Grid container spacing={1}>
 
                         <Grid item xs={1} />
                         <Grid item xs={1}>
