@@ -1,25 +1,28 @@
 import React from 'react';
+import 'babel-polyfill'
 import ReactDOM from 'react-dom'
-import { createStore } from 'redux'
 import { Provider } from 'react-redux'
 import { HashRouter } from 'react-router-dom'
 import { Route } from 'react-router'
 
-import { SignIn } from './auth/sign_in'
-import { ResetPasswordRequest } from './auth/reset_password_request'
-import { ResetPasswordResponse } from './auth/reset_password_response'
-import { OpenSourceLicenses } from './auth/open_source_licenses'
-import { LicenseAgreement } from "./auth/license_agreement";
+import configureStore from "./reducers/configureStore";
+import {saveState} from "./reducers/stateLoader";
 
-import { Home } from './kb/home'
+// must not have {} to work
+import SignIn from './auth/sign_in'
 
-export const sessionReducer = (state = {}) => {
-    return state;
-};
+import ResetPasswordRequest from './auth/reset_password_request'
+import ResetPasswordResponse from './auth/reset_password_response'
+import OpenSourceLicenses from './auth/open_source_licenses'
+import LicenseAgreement from "./auth/license_agreement";
 
-export let store = createStore(
-    sessionReducer
-);
+import Home from './home'
+
+const store = configureStore(history);
+store.subscribe(() => {
+    saveState(store.getState());
+});
+
 
 ReactDOM.render(
     <Provider store={store}>
@@ -28,8 +31,6 @@ ReactDOM.render(
             <div>
                 <Route exact path="/" component={SignIn} />
                 <Route exact path="/home" component={Home} />
-                <Route exact path="/sign-in" component={SignIn} />
-                {/*<Route exact path="/register" component={Register} />*/}
                 <Route path="/reset-password-request" component={ResetPasswordRequest} />
                 <Route path="/reset-password-response" component={ResetPasswordResponse} />
                 <Route exact path="/os-license" component={OpenSourceLicenses} />
