@@ -1,13 +1,13 @@
 
 const system_config = {
     // SimSage platform version (used for display only in UI)
-    version: '5.0.16',
+    version: '5.0.18',
     // api version of api_base
     api_version: 1,
     // the service layer end-point, change "localhost:8080" to ...
     api_base: 'http://localhost:8080/api',
     // the web-site to go to when SimSage logo is clicked on the login page
-    web_base: 'http://simsage.nz/',
+    web_base: 'https://simsage.nz/',
     // date picker display format
     date_format: 'yyyy/MM',
 };
