@@ -15,10 +15,10 @@ import TextField from '@material-ui/core/TextField';
 import Chip from '@material-ui/core/Chip';
 import TablePagination from '@material-ui/core/TablePagination';
 
-import {Api} from '../common/api'
-import {Comms} from '../common/comms'
-import {MessageDialog} from '../common/message-dialog'
-import {ErrorDialog} from '../common/error-dialog'
+import {Home} from '../home';
+import {connect} from "react-redux";
+import {bindActionCreators} from "redux";
+import {appCreators} from "../actions/appActions";
 
 
 const styles = {
@@ -35,6 +35,11 @@ const styles = {
         padding: '10px',
         color: '#888',
         cursor: 'pointer',
+    },
+    disabledLinkButton: {
+        float: 'left',
+        padding: '10px',
+        color: '#eee',
     },
     imageButton: {
         float: 'right',
@@ -70,6 +75,10 @@ const styles = {
     dlImageSize: {
         width: '24px',
     },
+    dlImageSizeDisabled: {
+        width: '24px',
+        opacity: '0.2'
+    },
     busy: {
         display: 'block',
         position: 'fixed',
@@ -85,145 +94,118 @@ const styles = {
     },
 };
 
-const roles = ['admin', 'operator', 'crawler', 'manager', 'reporter'];
+const roles = ['admin', 'operator', 'crawler', 'manager']; // reporter has been removed temporarily
 
 export class UserManager extends React.Component {
     constructor(props) {
         super(props);
-        this.kba = props.kba;
         this.state = {
-            has_error: false,
-            onError : props.onError,
-
-            busy: false,
-
-            user_list: [],
-
             edit_user: false,
             user: null,
-
+            edit_user_id: "",
             edit_first_name: "",
             edit_surname: "",
             edit_email: "",
             edit_password: "",
             edit_roles: [],
             edit_kb_list: [],
-
-            error_msg: "",
-            error_title: "",
-
-            message_title: "",
-            message: "",
-
-            openDialog: props.openDialog,
-            closeDialog: props.closeDialog,
-
             // pagination
             page_size: 5,
             page: 0,
-
-            selected_organisation_id: props.selected_organisation_id,
         };
     }
     componentDidCatch(error, info) {
-        this.setState({ has_error: true });
+        this.props.setError(error, info);
         console.log(error, info);
     }
-    componentWillReceiveProps(props) {
-        this.kba = props.kba;
-        const prev_selected_organisation_id = this.state.selected_organisation_id;
-        this.setState({
-            selected_organisation_id: props.selected_organisation_id,
-            openDialog: props.openDialog,
-            closeDialog: props.closeDialog,
-        });
-        if (prev_selected_organisation_id !== props.selected_organisation_id) {
-            this.refreshUsers(props.selected_organisation_id);
+    numUsers(organisation_id, isAdmin) {
+        let count = 0;
+        for (const i in this.props.user_list) {
+            const user = this.props.user_list[i];
+            const roleStr = UserManager.formatRoles(this.props.selected_organisation_id, user.roles);
+            if (isAdmin || roleStr.length > 0) { // has a role or isAdmin?
+                count += 1;
+            }
         }
+        return count;
     }
-    componentDidMount() {
-        this.refreshUsers(this.state.selected_organisation_id);
-    }
-    refreshUsers(org_id) {
-        if (org_id && org_id.length > 0) {
-            this.setState({busy: true});
-            Api.getUsers(org_id,
-                (userList) => {
-                    this.setState({user_list: userList, busy: false})
-                },
-                (errStr) => {
-                    this.showError("Error", errStr)
-                }
-            )
-        }
-    }
-    changePage(page) {
-        this.setState({page: page});
-    }
-    getUsers() {
+    getUsers(isAdmin) {
         const paginated_list = [];
         const first = this.state.page * this.state.page_size;
         const last = first + this.state.page_size;
-        for (const i in this.state.user_list) {
-            if (i >= first && i < last) {
-                paginated_list.push(this.state.user_list[i]);
+        let index = 0;
+        for (const i in this.props.user_list) {
+            // paginate all users - but only those that have roles in this organisation
+            const user = this.props.user_list[i];
+            const roleStr = UserManager.formatRoles(this.props.selected_organisation_id, user.roles);
+            if (isAdmin || roleStr.length > 0) { // has a role or is admin?
+                if (index >= first && index < last) {
+                    paginated_list.push(user);
+                }
+                index += 1; // one more user in this set of roles
             }
         }
         return paginated_list;
     }
+    changePage(page) {
+        this.setState({page: page});
+    }
     changePageSize(page_size) {
         this.setState({page_size: page_size});
-    }
-    showError(title, error_msg) {
-        this.setState({error_title: title, error_msg: error_msg, busy: false});
-    }
-    closeError() {
-        this.setState({error_msg: ''});
     }
     static formatRoles(organisationId, roles) {
         let roles_str = "";
         for (const role of roles) {
-            if (role.organisationId === organisationId) {
-                if (roles_str.length > 0) {
-                    roles_str += ", ";
+            // admin always displays
+            if (role.organisationId === organisationId || role.role === "admin") {
+                // make sure we add the admin role only once
+                if (role.role === "admin" && roles_str.indexOf("admin") === -1) {
+                    if (roles_str.length > 0) {
+                        roles_str += ", ";
+                    }
+                    roles_str += role.role;
+                } else if (role.role !== "admin") {  // any other role just add
+                    if (roles_str.length > 0) {
+                        roles_str += ", ";
+                    }
+                    roles_str += role.role;
                 }
-                roles_str += role.role;
             }
         }
         return roles_str;
     }
     getAvailableRoles() {
         const list = [];
-        for (const avaible_role of roles) {
+        const isAdmin = Home.hasRole(this.props.user, ['admin']);
+        for (const available_role of roles) {
             let found = false;
             for (const role of this.state.edit_roles) {
-                if (avaible_role === role) {
+                if (available_role === role) {
                     found = true;
                 }
             }
             if (!found) {
-                list.push(avaible_role);
+                if (available_role === 'admin' && isAdmin) {
+                    list.push(available_role);
+                } else if (available_role !== 'admin') {
+                    list.push(available_role);
+                }
             }
         }
         return list;
     }
     addNewUser() {
-        if (this.state.openDialog) {
-            this.state.openDialog();
-        }
-        this.setState({edit_user: true, knowledgeBase: null,
-            edit_kb_id: "",
-            edit_first_name: "",
-            edit_surname: "",
-            edit_email: "",
-            edit_password: "",
-            edit_roles: [],
-            edit_kb_list: [],
-        })
+        this.setState({edit_user: true, user: null,
+                            edit_user_id: "", edit_first_name: "",
+                            edit_surname: "", edit_email: "",
+                            edit_password: "",
+                            edit_roles: [],
+                            edit_kb_list: [],
+                        })
     }
     editUser(user) {
         if (user) {
-            const filtered_roles = user.roles.filter(x => x.organisationId === this.state.selected_organisation_id);
+            const filtered_roles = user.roles.filter(x => x.organisationId === this.props.selected_organisation_id);
             const role_list = [];
             for (const role of filtered_roles) {
                 role_list.push(role.role);
@@ -231,18 +213,15 @@ export class UserManager extends React.Component {
             const kb_list = [];
             if (user.operatorKBList) {
                 for (const operatorKB of user.operatorKBList) {
-                    for (const available_kb of this.kba.knowledge_base_list) {
+                    for (const available_kb of this.props.knowledge_base_list) {
                         if (available_kb.kbId === operatorKB.kbId) {
                             kb_list.push(available_kb);
                         }
                     }
                 }
             }
-            if (this.state.openDialog) {
-                this.state.openDialog();
-            }
-            this.setState({edit_user: true, knowledgeBase: user,
-                edit_kb_id: user.id,
+            this.setState({edit_user: true, user: user,
+                edit_user_id: user.id,
                 edit_first_name: user.firstName,
                 edit_surname: user.surname,
                 edit_email: user.email,
@@ -297,27 +276,21 @@ export class UserManager extends React.Component {
     }
     deleteUserAsk(user) {
         if (user) {
-            this.setState({message_title: "Remove User from Organisation",
-                message: "are you sure you want to remove " + user.firstName + " " + user.surname + " from this organisation?",
-                knowledgeBase: user})
+            this.props.openDialog("are you sure you want to remove " + user.firstName + " " + user.surname + " from this organisation?",
+                                  "Remove User", (action) => { this.deleteUser(action) });
+            this.setState({user: user});
         }
     }
     deleteUser(action) {
         if (action) {
-            this.setState({busy: true});
-            Api.removeUserFromOrganisation(Comms.getSession().userId, this.state.selected_organisation_id, () => {
-                this.setState({message_title: "", message: "", busy: false});
-                this.refreshUsers(this.state.selected_organisation_id);
-            }, (errStr) => {
-                this.setState({message_title: "", message: "", busy: false,
-                                     error_msg: errStr, error_title: "Error Removing User"});
-            })
-        } else {
-            this.setState({message_title: "", message: "", busy: false});
+            this.props.deleteUser(this.props.selected_organisation_id, this.state.user.id);
+        }
+        if (this.props.closeDialog) {
+            this.props.closeDialog();
         }
     }
     editCancel() {
-        this.setState({edit_user: false, knowledgeBase: null})
+        this.setState({edit_user: false, user: null})
         if (this.state.closeDialog) {
             this.state.closeDialog();
         }
@@ -326,42 +299,40 @@ export class UserManager extends React.Component {
         if (this.state.edit_first_name.length > 0 &&
             this.state.edit_surname.length > 0 &&
             this.state.edit_email.length > 0 &&
-            (this.state.edit_kb_id.length > 0 || this.state.edit_password.length > 0) &&
+            (this.state.edit_user_id.length > 0 || this.state.edit_password.length > 0) &&
             this.state.edit_roles.length > 0) {
 
             // do we have an operator and knowledge bases?
             if (UserManager.hasOperatorRole(this.state.edit_roles) && this.state.edit_kb_list.length === 0) {
-                this.setState({
-                    error_msg: "An operator must have at least one knowledge base associated with it.",
-                    error_title: "Incomplete Data"});
-
+                this.props.setError("Incomplete Data", "An operator must have at least one knowledge base associated with it.");
             } else {
-                this.setState({busy: true});
-                Api.updateUser(this.state.selected_organisation_id, this.state.edit_kb_id,
-                    this.state.edit_first_name, this.state.edit_surname, this.state.edit_email,
-                    this.state.edit_password, this.state.edit_roles, this.state.edit_kb_list,
-                    (user) => {
-                        this.setState({edit_user: false, knowledgeBase: null, busy: false});
-                        this.refreshUsers(this.state.selected_organisation_id);
-                        if (this.state.closeDialog) {
-                            this.state.closeDialog();
-                        }
-                    },
-                    (errStr) => {
-                        this.setState({edit_user: false, error_msg: errStr, error_title: "Error Updating User", busy: false});
-                        if (this.state.closeDialog) {
-                            this.state.closeDialog();
-                        }
-                    });
+                // organisation_id, user_id, name, surname, email, password, role_list, kb_list
+                this.props.updateUser(this.props.selected_organisation_id, this.state.edit_user_id,
+                                      this.state.edit_name, this.state.edit_surname, this.state.edit_email,
+                                      this.state.edit_password, this.state.edit_roles, this.state.edit_kb_list);
+                this.setState({edit_user: false, user: null});
             }
-
         } else {
-            this.setState({
-                error_msg: "Please complete all fields.  " +
-                    "Must have email-address, first-name, surname, and at least one role.  " +
-                    "New Accounts must have a password.",
-                busy: false,
-                error_title: "Incomplete Data"});
+            this.props.setError("Incomplete Data", "Please complete all fields.  " +
+                "Must have email-address, first-name, surname, and at least one role.  " +
+                "New Accounts must have a password.");
+        const self = this;
+        this.setState({message: '', message_title: ''});
+        if (confirm) {
+            this.setState({busy: true});
+            const crawler = this.state.crawler_ask;
+            const self = this;
+            Api.deleteCrawler(this.props.selected_organisation_id, this.props.selected_knowledgebase_id, crawler.id,
+                (response) => {
+                    this.setState({busy: false});
+                    self.getCrawlerList(self.state.page, self.state.page_size);
+                },
+                (error) => {
+                    self.props.setError('Error', error);
+                }
+            );
+        }
+
         }
     }
     static hasOperatorRole(edit_roles) {
@@ -390,7 +361,7 @@ export class UserManager extends React.Component {
     }
     getAvailableKnowledgeBases() {
         const list = [];
-        for (const available_kb of this.kba.knowledge_base_list) {
+        for (const available_kb of this.props.knowledge_base_list) {
             let found = false;
             for (const kb of this.state.edit_kb_list) {
                 if (available_kb.kbId === kb.kbId) {
@@ -404,24 +375,11 @@ export class UserManager extends React.Component {
         return list;
     }
     render() {
-        if (this.state.has_error) {
-            return <h1>users.js: Something went wrong.</h1>;
-        }
+        const isAdmin = Home.hasRole(this.props.user, ['admin']);
         return (
             <div>
-                <ErrorDialog
-                    callback={() => { this.closeError() }}
-                    open={this.state.error_msg.length > 0}
-                    message={this.state.error_msg}
-                    title={this.state.error_title} />
-
-                <MessageDialog callback={(action) => { this.deleteUser(action) }}
-                    open={this.state.message.length > 0}
-                    message={this.state.message}
-                    title={this.state.message_title} />
-
                 {
-                    this.state.busy &&
+                    this.props.busy &&
                     <div style={styles.busy} />
                 }
 
@@ -438,7 +396,7 @@ export class UserManager extends React.Component {
                         </TableHead>
                         <TableBody>
                             {
-                                this.getUsers().map((user) => {
+                                this.getUsers(isAdmin).map((user) => {
                                     return (
                                         <TableRow key={user.id}>
                                             <TableCell>
@@ -451,15 +409,37 @@ export class UserManager extends React.Component {
                                                 <div style={styles.label}>{user.surname}</div>
                                             </TableCell>
                                             <TableCell>
-                                                <div style={styles.label}>{UserManager.formatRoles(this.state.selected_organisation_id, user.roles)}</div>
+                                                <div style={styles.label}>{UserManager.formatRoles(this.props.selected_organisation_id, user.roles)}</div>
                                             </TableCell>
                                             <TableCell>
-                                                <div style={styles.linkButton} title="Edit this user" onClick={() => this.editUser(user)}>
-                                                    <img src="../images/edit.svg" style={styles.dlImageSize} title="edit user" alt="edit"/>
-                                                </div>
-                                                <div style={styles.linkButton} title="Remove this user" onClick={() => this.deleteUserAsk(user)}>
-                                                    <img src="../images/delete.svg" style={styles.dlImageSize} title="remove user" alt="remove"/>
-                                                </div>
+                                                { Home.hasRole(user, ['admin']) && !isAdmin &&
+                                                    <div style={styles.disabledLinkButton} title="cannot edit this user">
+                                                        <img src="../images/edit.svg" style={styles.dlImageSizeDisabled}
+                                                             title="cannot edit this user" alt="cannot edit this user"/>
+                                                    </div>
+                                                }
+                                                { (isAdmin || !Home.hasRole(user, ['admin'])) &&
+                                                    <div style={styles.linkButton} title="Edit this user"
+                                                         onClick={() => this.editUser(user)}>
+                                                        <img src="../images/edit.svg" style={styles.dlImageSize}
+                                                             title="edit user" alt="edit"/>
+                                                    </div>
+                                                }
+                                                {
+                                                    (this.props.user.email === user.email || !isAdmin) &&
+                                                    <div style={styles.disabledLinkButton} title="cannot remove this user">
+                                                        <img src="../images/delete.svg" style={styles.dlImageSizeDisabled}
+                                                             title="cannot remove this user" alt="cannot remove this user"/>
+                                                    </div>
+                                                }
+                                                {
+                                                    this.props.user.email !== user.email && isAdmin &&
+                                                    <div style={styles.linkButton} title="Remove this user"
+                                                         onClick={() => this.deleteUserAsk(user)}>
+                                                        <img src="../images/delete.svg" style={styles.dlImageSize}
+                                                             title="remove user" alt="remove"/>
+                                                    </div>
+                                                }
                                             </TableCell>
                                         </TableRow>
                                     )
@@ -471,31 +451,34 @@ export class UserManager extends React.Component {
                                 <TableCell />
                                 <TableCell />
                                 <TableCell>
-                                    {this.state.selected_organisation_id.length > 0 &&
+                                    {this.props.selected_organisation_id.length > 0 && isAdmin &&
                                     <a style={styles.imageButton} onClick={() => this.addNewUser()}><img
                                         style={styles.addImage} src="../images/add.svg" title="add new user"
                                         alt="add new user"/></a>
                                     }
                                 </TableCell>
                             </TableRow>
+                            <TableRow>
+                                <TableCell colSpan={5}>
+                                    <TablePagination
+                                        rowsPerPageOptions={[5, 10, 25]}
+                                        component="div"
+                                        count={this.numUsers(this.props.selected_organisation_id, isAdmin)}
+                                        rowsPerPage={this.state.page_size}
+                                        page={this.state.page}
+                                        backIconButtonProps={{
+                                            'aria-label': 'Previous Page',
+                                        }}
+                                        nextIconButtonProps={{
+                                            'aria-label': 'Next Page',
+                                        }}
+                                        onChangePage={(event, page) => this.changePage(page)}
+                                        onChangeRowsPerPage={(event) => this.changePageSize(event.target.value)}
+                                    />
+                                </TableCell>
+                            </TableRow>
                         </TableBody>
                     </Table>
-
-                    <TablePagination
-                        rowsPerPageOptions={[5, 10, 25]}
-                        component="div"
-                        count={this.state.user_list.length}
-                        rowsPerPage={this.state.page_size}
-                        page={this.state.page}
-                        backIconButtonProps={{
-                            'aria-label': 'Previous Page',
-                        }}
-                        nextIconButtonProps={{
-                            'aria-label': 'Next Page',
-                        }}
-                        onChangePage={(event, page) => this.changePage(page)}
-                        onChangeRowsPerPage={(event) => this.changePageSize(event.target.value)}
-                    />
 
                 </Paper>
 
@@ -507,8 +490,8 @@ export class UserManager extends React.Component {
                         disableEscapeKeyDown={true}
                         fullWidth={true}
                         maxWidth="md"
-                        onClose={() => this.setState({edit_user: false, knowledgeBase: null})} >
-                    <DialogTitle>{this.state.edit_kb_id ? "Edit User" : "Add New User"}</DialogTitle>
+                        onClose={() => this.setState({edit_user: false, user: null})} >
+                    <DialogTitle>{this.state.edit_user_id ? "Edit User" : "Add New User"}</DialogTitle>
                     <DialogContent>
                         <TextField
                             autoFocus={true}
@@ -612,4 +595,21 @@ export class UserManager extends React.Component {
     }
 }
 
-export default UserManager;
+const mapStateToProps = function(state) {
+    return {
+        error: state.appReducer.error,
+        error_title: state.appReducer.error_title,
+
+        user_list: state.appReducer.user_list,
+        selected_organisation_id: state.appReducer.selected_organisation_id,
+        knowledge_base_list: state.appReducer.knowledge_base_list,
+        busy: state.appReducer.busy,
+        user: state.appReducer.user,
+    };
+};
+
+export default connect(
+    mapStateToProps,
+    dispatch => bindActionCreators(appCreators, dispatch)
+)(UserManager);
+
